refactoring to make the resource monitor code more unitised and unit tested
